requirements.txt
streamlit
faiss-cpu
sentence-transformers
PyMuPDF
tqdm
fitz
pdfplumber
pytesseract
pdf2image
